Simple RSA model Kullback-Leibler divergence determination (available feature values only)

Simple RSA

The function calculates the optimal parameter values of the free parameters by estimating the log-likelihood of the RSA model given model parameters and data. It also determines the actual RSA model Kullback-Leibler divergence.

3 parameter optimization considering only the available feature values present in the scene, i.e. feature values of shape, texture and color.

RSAModelKLDiv3paramsOnlyAvailableFeatureValuesConsidered_simpleRSA(
  data,
  par1,
  par2
)

Arguments

data
A matrix with data rows.

column structure:

[1:OC1,OC2,OC3,4:UUFeat, 5:Q1Feat,6:Q2Feat]

[7:Q1AnswerV1,V2,V3, 10:Q2AnswerV1,V2,V3]

1:OC1 Object 1. A value between 1 and 27.

2:OC2 Object 2. A value between 1 and 27.

3:OC3 Object 3. A value between 1 and 27.

4:UUFeat Uttered feature. A number between 1 and 3. (1: shape, 2: pattern, 3: color)

5:Q1Feat Questioned feature 1. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

6:Q2Feat Questioned feature 2. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

7:Q1AnswerV1, V2, V3 The columns 7-9 contain the participants' slider values for the first questioned feature.

10:Q2AnswerV1, V2, C3 The columns 10-12 contain the participants' slider values for the second questioned feature.

par1

softness parameter

The strength of "preferring one entity over others". (The larger the value the higher the tendency towards uniform liking)

par2

non-obedience parameter

The extent to which the instruction of the speaker is obeyed by the listener. (0 = full obedience, infinity = full instruction ignorance).

Value

Minimized Kullback-Leibler divergence and the optimal parameters.

Details

This function is used in RSAModelLL1_1simpleRSA, RSAModelLL1_1simpleRSA_notObey.1,

RSAModelLL1_1simpleRSA_notObey.2, RSAModelLL1_2simpleRSA,

RSAModelLL1_2simpleRSA_pref2, RSAModelLL2_simpleRSA.