Experiment 2 raw data (Epistemic utterance choice)

This file contains the raw data from the utterance choice experiment.

X3_rawData

Format

A data frame with the following columns:

workerid

Participant's ID

item

Unique ID of a stimulus

slide_number

Indicates trial order.

condition

A six-digit code that infers to which category the item belongs. In this experiment, it is only necessary to identify classes that have potentially ambiguous utterances.

language

Participant self-reported native language

pref1-pref9

Utterance 1-9.

response1- response9

The value to which the slider (1-9) was adjusted by the participant. Indicates how useful an utterance is for finding out listener preferences. The corresponding property that they are rating is stored under pref1-9. Can be in range 0-1. Small numbers mean less useful, larger numbers mean more useful.

obj1, obj2, obj3

A three-digit code referring to the object that the speaker intends to signal by the utterance (obj1), or to one of the other two objects present in the scene (obj2, obj3). The first digit of the code refers to shape, the second digit to texture, and the third digit to color of the object. 1 Shape: 1 cloud, 2 circle, 3 square 2 Texture: 1 solid, 2 striped, 3 polka-dotted 3 Color: 1 blue, 2 red, 3 green Example: If the target is a solid blue square the code would be: 311.

ambiguous

Marks whether a trial contains any ambiguous utterances.

numFeatures

The number of feature values present in a scene. For example, if we have 3 red objects, all of them solid, 1 square, 1 cloud, 1 circle, then the number of features would be 5. Value in range 3-9.