Get priors of object preferences of the listener

Determine the priors of object preferences of the listener.

These are automatically derived from the valid utterances in the scene (i.e. derived from all features of the current objects).

getObjectPreferencePriors(validUtterances, currentObjects, type, uttToObjProbs)

Arguments

validUtterances
A vector of utterances that correspond to all feature values present in the current objects in the scene.

For example, it only makes sense to utter "red" in a scene if there are red objects present.

currentObjects
A vector of three values in range {1,...,27} specifying the target and the other two objectsin the scene.

The target is the first object in the vector (index = 1).

type
Is set to 0: Hard priors or >0: Soft priors with specified softness

uttToObjProbs
A matrix. The rows map each possible utterance that corresponds to each present feature value of the current objects. The columns represent the three objects in the scene.

This reflects the obedience-parameter and which objects match the respective utterance. The matrix shows the probability that a certain object is chosen following a certain utterance, that is valid in the scene. The number of rows of the matrix match the length of the validUtterances vector.

Value

A list of preference priors for all valid utterances based on the objects in the scene.

The list has as many rows as the length of the validUtterances vector + 1.

Each row in the list contains a vector of length 3, as there are three objects in the scene.

The extra row is for the case of no preferences whatsoever, i.e. uniform prior over all three objects in the scene.

Details

This function is only used in X9

Examples