Get the utterances for the speaker's best information gain

Simple RSA

Determines the optimal utterance for the best information gain.

These are based on the valid utterances determined from the current objects in the scene.

The inferred listener's object preferences are computed assuming the listener picks a certain object and has certain object preferences.

getSimpleBestInfGainUttPreferences(
  currentObjects,
  softPrefValue,
  notObeyInst,
  klValueFactor
)

Arguments

currentObjects
A vector of three values in {1,...,27} specifying the target and the other two objects in the scene.

The target is the first object in the vector (index = 1).

softPrefValue
A parameter value between [0,infinity) (The larger the value the higher the tendency towards uniform liking).

Value reflects how categorical the listener's preferences are:

0: The listener always picks her preferred object.

If the listener prefers red objects, she will always pick the red object in the scene.

infinity: It is as likely for the listener to pick green, blue or red objects.

notObeyInst
Determines the extent to which the instruction of the speaker is obeyed by the listener.

(0 = full obedience, infinity = full instruction ignorance).

Example:

0: Listener always picks red objects following the utterance "red".

infinity: Listener as likely to pick green, blue or red objects even if the utterance is "red".

klValueFactor
A parameter that can be negative, zero or positive:

zero

Don't care about learning about feature preferences of the listener

positive

Care about learning about feature preferences of the listener

negative

Trying to pick non-ambiguous utterances

Value

A vector containing the normalized probability over utterances given the listener's object preference priors.

The utterance with the highest probability is the one that maximizes the information gain for the speaker.

The vector has the same length as the validUtterances vector.

Details

This function uses the function simpleBestInfGainUtterance.

Examples